fix universe map in infcx.instantiate_canonical_*?
Previously `infcx.instantiate_canonical_with_fresh_inference_vars` maps the root universe in `canonical` to `ty::UniverseIndex::ROOT`, assuming it's working with a fresh InferCtxt, but that's not true for the use cases in mir typeck. No the root universe is mapped into `infcx.universe()`.

I'm not sure if this  is the right fix or if it's really a bug, but it looks wrong to me!